requirements.txt
Django==2.2.16
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.1
pyflakes==2.5.0
PyJWT==2.5.0
Pillow==9.2.0
djoser==2.1.0